✨ 키워드 스타일 선택, 이미지 생성 api 구현
