+++ aif_group/src/hooks/useImages.ts
@@ -1,3 +1,4 @@
+import { createImages } from '@/services/createImages';
 import { getImages } from '@/services/getImages';
 import { useQuery } from '@tanstack/react-query';
 
@@ -5,13 +6,15 @@ const email = 'b4ur2old@naver.com';
 export const useImages = (keyword: string, style: string) => {
   const { data, isLoading, error, refetch } = useQuery({
     queryKey: ['getImageDataApi', email],
-    queryFn: () =>
-      getImages()
-        .then(res => res)
-        .catch(err => {
-          console.error(err);
-          return [];
-        }),
+    queryFn: async () => {
+      try {
+        await createImages(keyword, style);
+        const ImageData = await getImages();
+        return ImageData;
+      } catch (error) {
+        console.error(error);
+      }
+    },
     enabled: false,
   });
   return { data, isLoading, error, refetch };
