+++ aif_group/src/hooks/useValidation.ts
@@ -1,0 +1,20 @@
+import { postValidation } from '@/services/postValidation';
+import { useQuery } from '@tanstack/react-query';
+
+function useValidation() {
+  const { isLoading, error, refetch } = useQuery({
+    queryKey: ['postValidation', email],
+    queryFn: async () => {
+      try {
+        const res = await postValidation(email);
+        console.log(res);
+      } catch (error) {
+        console.error(error);
+      }
+    },
+    enabled: false,
+  });
+  return { isLoading, error, refetch };
+}
+
+export default useValidation;
